fix(backend): not found on user delete

+++ backend/src/users/users.service.ts
@@ -69,6 +69,11 @@ export class UsersService {
   // }
 
   async remove(id: number): Promise<void> {
+    const user = await this.userRepository.findOneBy({ id })
+    if (!user) {
+      throw new NotFoundException('User not found')
+    }
+
     await this.userRepository.delete(id)
   }
 }
